pkg/auth/badgerauth: bump MaximumBufferSize within DRPC for nodes
Sometimes we have to put machines where nodes reside into maintenance
and disconnect from the rest of the cluster. This means they will get
desynced. We observed that if a node gets desynced from all other nodes,
even though we limit the number of records ("replication limit") other
nodes can replicate to the desynced node in a single batch, it's still
more data than what nodes are allowed to read in one go (4 MiB). This
causes replication failures and effectively disables the desynced node
from syncing again until manually intervening. We could lower the
replication limit further to allow data to pass in fewer batches, but we
still want to keep the replication swift, which means bumping the amount
nodes can read is the left option. This change bumps the limit 10x (to
40 MiB).

Change-Id: I816e1ee3406f80fbcc5e513b433ae0f636a89f74