cmd/authservice-admin: add links commands to inspect and revoke links
This adds the following new commands to authservice-admin:
* links inspect
* links revoke

These commands can be used to get details on one or more links, without
the need to manually check each one. It can also be used to revoke
access to one or more links at once.

Additionally, this also changes the existing record commands to
optionally get more details on the API key from the satellite admin if
it's configured. If it's not configured, it just uses the authservice
nodes only.

It also adjusts the "tabbed" output of record commands to be consistent
with these new links commands and list out each field on a new line, as
the existing tabbedwriter output was becoming very difficult to read
with so many columns.

Updates #321

Change-Id: I9d0bc1c73c389384e05e729eac3449de665292af