pkg/auth/badgerauth: add a timeout for replication
In badgerauth's code, we had a timeout for everything but the sync
operation. We never added it because we assumed that the operation would
eventually succeed. Unfortunately, this isn't always the case. Nodes
would establish connections between themselves, but the harsh reality of
Internet networks is that frames and packets can get lost along the way,
and we would block the sync forever, waiting for the response from the
other side of the connection that might be long gone. Whether this is
related to networking misconfiguration or DRPC's handling of
broken/stale connections (we haven't really found compelling evidence
that DRPC's behaving in a way it's not supposed to), the priority is to
keep sync running. This change caps the time to dial the connection and
replicate from all peers at (number of peers) * (replication interval)
(instead of just (replication interval) to make sure we have some room
for (un)expected latency there).

Updates storj/storj-private#284

Change-Id: I2be1d98edfa01cd04bff11dbe59f9be92f4986bc