internal/authadminclient: allow host in authnode addresses
when authservice-admin is being run on local machines, there may not
DNS entries to resolve auth node hostnames to IP addresses without
the need to add /etc/hosts entries yourself.

hostname is needed when dialing to successfully do certificate
authentication with auth nodes, so we can't use IP addresses directly.

this now allows for either format to be set for node addresses:

* authnode-dc1-app1:20004
* 1.2.3.4:20004=authnode-dc1-app1

Updates #321

Change-Id: Ie7c7d324416cd78e7c9670458f7d859e7c45fe1a