cmd/authservice-admin: allow disabling api key deletion if requested
if we're deleting records on authservice and satellite admin addresses
are configured, the command will also attempt to delete the api key
on the satellite.

the key may have already been deleted on the satellite, so the command
will error that it can't find the key. This adds the ability to use
--delete-api-key=false so it doesn't try to delete the key on the
satellite, and only delete the authservice node records only.

Change-Id: Icaaf4820f0bd8d25e31faca31110bb2c03aaf916