internal/authadminclient: report all errors when modifying records
currently we return the first error when calling out to all authservice
nodes to modify or delete a record, but since only the first non-nil
error is returned we don't know which other nodes it may have
encountered an error on.

this now returns a combined error of every node which returned a
non-nil error.

Change-Id: I4ddd9b611f55a77b9bc8f515071524c00c2e54c3