cmd/authservice-admin: ensure we always attempt to delete access keys
if revoking a set of links that share the same API key we might find
that some results show an error like "api key not found" but don't end
up deleting the authservice records specific to that link.

this changes the revoke command to delete api key and authservice
records at the same time, and only attempt to delete the api key if it
actually exists.

Change-Id: If7aeec468a59068c65db4256b4d9c6cc4b72997e