pkg/auth: make certificate setup on-demand to fix getting first cert
it's currently not possible to get a certificate on first run of
authservice because certmagic is trying to use TLS-ALPN challenge
for a TLS listener that hasn't started yet.

to avoid that situation we're setting OnDemand config which delays
the certificate provisioning until later so we can successfully
retrieve the certificate after the listener started. Linksharing
already does this.

Change-Id: I155162bbad861683a7a09f8bf3f7a6140de03141