pkg/linksharing/sharing: alter response for bandwidth limit error
this changes the HTTP status for exceeding bandwidth to 403, to be
consistent with gateway.

429 status should only be used for request rate limits, but not
other limits as it's not semantically correct. Storage and segments
limits don't need to be checked, as those only kick in when you
attempt to upload an object, but Linksharing has no such
capability.

this also includes a minor adjustment in response text for other
errors to make them consistent and more to the point.

Change-Id: I820d597e4a8d5a18567e3290241f7246437d50c4