{pkg/linksharing,testsuite}: fix landing page redirect
regression in credential middleware meant that an empty path was
not being redirected to the landing page target and instead tried
to do a backwards compatible link redirect. This meant clicking
the "Go to homepage" button would redirect to
"https://link.storjshare.io/s" and show a malformed request error.

Change-Id: I812ab8ee1d3014b58d31614f4dd56134502d9461