linksharing: list forward pagination
Before this change we would list every single path under the prefix. If
there were a lot of paths, this could perform poorly. Now there is a
linksharing config --list-page-limit which will limit the number of
objects we display on the page. A query parameter, cursor, has been
added to the endpoint, which is used as the starting point for a list
request. On a list request if there are more objects than the limit, a
button is displayed which uses the last object in the current batch as
the cursor for the next request.

As we only list forwards, implementing the ability to go back to
previous pages requires listing from the beginning and keeping track of
the cursor for the previous page until we reach the page we want to
display. This is proving to be more complex, so we can work on it
separately as this commit already solves the main concern with listing.

closes #309

Change-Id: I3ca58d428fe2155ea5dd027902201c9f33b6c7d2