linksharing: list prefix back buttons
This change implements a button to backwards to the previous page in the
browser history and a button to go back to the first page of the prefix
list. The buttons are only displayed if the cursor parameter is set,
which means we are not on the first page. The back button is limited in
that if the user navigates directly to a linksharing page without having
navigated from a previous linksharing page, the back button will take
them back to whatever non-linksharing page they came from. This is the
purpose of the "Back To Page 1" button. We implemented the "Back"
functionality this way due to the complexity of keeping track of the
cursor of the previous page, as we currently can only list forwards.
However, we may implement reverse listing in the future.

Change-Id: I30a353cc5d8f38d0776543f88438994058506e68