pkg/auth/spannerauth: fix zero time for expires_at records
currently we are setting a zero time value in the database
during the spannerauthmigration process, e.g.
0001-01-01T00:00:00Z, which results in row deletion policy / TTL
deleting records that weren't supposed to expire.

this change adds a check to spannerauth.Put to ensure we aren't
going to write a zero value expiry time, but keep it as null so
the row deletion policy won't delete them.

Change-Id: I15d289b9db79ce6573db9b3f27cc301f2d1eae24