pkg/auth/badgerauth: don't return invalid records from peers
When a record is requested from a badgerauth node, it first checks if
it has the record, and if it doesn't, it asks its peers for it.
The method it uses to read from itself is required to return an error
if the record has been invalidated, but the method for reading from
peers has no such constraint. This made it possible for a node to
return invalid records from its peers. This change fixes this issue.

Change-Id: I2897877f2a262002a782b9debeaeebde1566fec7