pkg/linksharing/sharing: download object prefix
If the `?download=1` or `wrap=0` query params are passed on an object
prefix, linksharing will stream recursively all objects with this prefix
as a zip file. Optionally, another query param `download-kind=tar.gz`
can be passed to download the same prefix as a .tar.gz file.

To avoid using too much memory when writing a zip file, a configurable
value is added to restrict downloading prefixes as zip with too many
files. This restriction does not apply to tar.gz files.

The full download prefix feature is behind a flag which is disabled by
default. Once the feature is considered production-safe, we can update
the flag's behavior or remove it entirely.

https://github.com/storj/storj-private/issues/833

Change-Id: I3ab11195f927edf8a09d0b3c7c034cf5de53295f