pkg/linksharing/sharing: don't list prefix file placeholder
When creating a folder in the satellite UI, we create an empty object
named ".file_placeholder" prefixed with the folder name. When listing a
shared prefix, this placeholder file shows up, and we don't want that.
To prevent showing this file, during the object iteration of a list
operation, skip it if it is found. Additionally, if there is a further
page, check if the file placeholder is the only object on it. If so,
treat the current page as the last page.

closes #6389

Change-Id: Id90146e4177a294cc102d1dbd27bda2ef9ec1e9c