pkg/auth/authdb: fix EqualWithinDuration to check within 1 minute
fix the case of authservice-admin getting
"mismatch between spanner and badger records" error when pulling
an expired record that exists in both badger and spanner backends.

badger currently returns ExpiresAt with seconds resolution, but
spannerauth is returning with nanoseconds resolution. We could fix
this, but it would require changing the protobuf, and since we
will switch entirely to spannerauth soon it seems easy enough to
just check ExpiresAt is roughly within a margin of error of 1 minute.

Change-Id: Ife50b86016847ba3fb5c73245ccf6ee68a24c3c4