internal/dbutil: Add internal dbutil package
common/dbutil was only depended on for a single function. Due to dbutil
moving to storj.io/storj/shared, we would need to introduce a dependency
to storj.io/storj in storj.io/edge, unless we remove the dbutil
dependency.

If we attempt to depend on storj.io/storj here, there are circular
dependency issues. For example, the storj.io/storj/testsuite project
depends on both storj.io/storj and storj.io/edge, which makes adding a
cross-dependency there difficult.

https://github.com/storj/storj/issues/6953

Change-Id: Ia5a7f338d4790d58ed3c82f4240a27c9a4c16b8e