pkg/server, cmd/linksharing: support specifying uplink client identities
this change allows the gateway and linksharing processes to specify
stable uplink client identities. this is important so that different
gateway regions can get different choice-of-two load balancing profiles,
and other benefits of satellite-trusted uplinks.

we intend that all gateways and linksharing services within a single
region use the same identity.

note that the linksharing process already has an identity called the
TierServiceIdentity, where it asks the Satellite for tier services using
that identity. If that identity is also different between regions and
the same within a region, it is fine to reuse that identity for this
as part of the configuration.

a future patch set could simplify the configuration so only one is
necessary to configure

Change-Id: If2942af0907126d1c91578d61af6cad43c13c69b