pkg/auth: save public project ID with authservice data
This change updates pkg/auth/authdb (*Database).Put() calls to
save public project ID with the access credentials in the
database. This works for both HTTP and DRPC ways of registering
credentials.

A flag is available to disable this if necessary. This is most
useful for unit tests that don't have a real or test satellite.

Updates #439

Change-Id: I1b074f226d8c7d4baa363373a5a5e2d65fa55646