pkg/accesslogs: add the package
This change adds a package that can handle collection and upload of
arbitrarily formatted server access logs in the fashion of S3's server
access logging.

API of the package is mostly just the collection engine (Processor) that
can be run and closed, queuing log entries upon request. The collection
engine works together with a concurrently running uploader tasked with
uploading to the Storage implementation. Because these concerns are
separated and of the way interfaces are designed, it allows for async
communication between the caller, the log processor and the uploader,
which makes queuing new log entries concise and fail only if certain
configurable limits are hit.

Logs are collected, packaged and uploaded when a certain size of the
package is hit. This can be later expanded to a time-based trigger.

Upon closing the peer, all buffers are flushed and the call to Close is
blocked until all flushing and uploading is done.

Current implementations of uploader and Storage are:

- sequentialUploader
    - uploads packages sequentially with configurable retries
- inMemoryStorage
    - useful in unit tests
- StorjStorage
    - implementation of Storage that writes to Storj using libuplink

What's left to implement to close the gap to production-ready server
access logging:

- (more) tests
    - for limits
    - E2E test(s) for sequentialUploader
    - E2E test(s) for Processor with StorjStorage
- benchmarks
- (Entry implementation) Amazon S3 server access log format
- wiring up Processor and entry queuing within the gateway

Closes storj/storj-private#824

Change-Id: I99578ac81f18f19c9eca2b256145bbe1493ca421