pkg/server/middleware: minor optimization to access log queueing
avoid the need to create the log entry and throw it away if there's
no logging configuration that matches the bucket and project ID
for the request. Additionally, expand test cases to ensure the
middleware behaves correctly on certain cases like malformed
project ID from authservice or it missing entirely.

Change-Id: Iff284f6d5c2b475e19c91b201fc297eb5b130286