pkg/auth/badgerauth: clean up the code after the purge
badgerauth is now just a simple implementation of Storage (and not
StorageAdmin until we need it) built on top of BadgerDB. This change
cleans up the remainder of what wasn't necessary for the "replication
purge" change. Documentation will be adjusted at a later time (I'm yet
not sure whether to get rid of it completely or adjust it to the current
situation).

Change-Id: I51c7de78616ea5c4c19fb2ac7ed85f0a45350733