pkg/server/gw: log errors in all cases
It would be more helpful for debugging customer issues if we
logged error messages in all cases, instead of not logging
expected errors. We are also seeing 500 status errors without
any error logged in the message.

Additionally, add a metric to see if we're not logging errors
in some cases. It's likely we'll want to clean this logging
code up soon, but this could be one more piece of evidence
that we're not logging properly in some cases.

Change-Id: I48eb757350ebc55062c0cd403f2e0a33216c4d21