pkg/linksharing/objectranger: avoid returning uplink.Download twice
Range function returns a io.ReadCloser, which will be closed by the caller.
Similarly, two range calls would lead them to share the download, which
is unsafe.

It does not seem we are using it in a context where Range is called in such
a way, but it's better to be safe than sorry.

Change-Id: I4b33ba3d71b40a5d2acc6123c0526231eb80268c