pkg/linksharing/{objectranger,sharing}: prevent double close
Download.Close was deferred in presentWithProject and then passed into
a ObjectRanger, which may be returned in httprange.ServeContent.
That Range result is closed in ServeContent, leading to a double close.

There is no easy way to currently propagate the close responsibility,
so instead, it adds a readCloserOnce, which then ensures that the
outer scope only closes the download when it has not been closed.

Change-Id: I2de9fdab9697cd79a0f77c5272802cfbe3e08611