cmd/authservice-admin: allow freezing unconditionally if requested
Now that we have restrictive time limits on free accounts, link
takedowns almost never freeze accounts anymore as it only happens
when the account is free.

This changes the behavior so if you set the flag to true
(default=false) it will freeze the account owner of the link(s)
that should be removed, regardless of tier status.

Change-Id: Ia23b1e4620141d6473130f813886051b7cb1a8f4