Jenkinsfile: don't run splunk, aws-sdk-go, and s3cmd tests in parallel
The splunk, aws-sdk-go, and s3cmd integration tests have been made to
run sequentially instead of in parallel with the other tests. This was
done to mitigate an issue that would occur because the satellite can't
handle certain parallel object deletions when the Spanner emulator is
used. These test suites are the ones that experience errors due to this
issue most often.

Change-Id: I6929431fb298f02ffcb4d902446621f53f0f9667