all: Homogenize zap fields
We have a disperse manners of writing the name of Zap logger field
names.

This commit homogenize all of them to use snake_case with only lowercase
ASCII letters, numbers and underscores.

This is a companion commit for the new linter that we'll detect Zap
field names which don't follow this convention
https: //review.dev.storj.tools/c/storj/ci/+/19816

Change-Id: Id3e33994d4dda3d5a0974d779521e8474d1a9bd5